An example about how to use reactive data to plot
This is due to that if we directly use reactive() to pass data to our plotting functions, we will encounter error like "Operation not allowed without an active reactive context. ". So we need to use isolate() function to create a non-reactive scope for an expression, then the problem can be perfectly solved.