Respond to the change in reactive data
Now we've found the doabler way to make our charts be able to respond to the change in the reactive data.

This is done together with observeEvent() function. Once there is any change in the data, the chart rendering function will be called again so that the chart can be updated according to the change in the data.