Modified the scatter render function (#13)
This commit is to better the axis range values display when turn off the auto.scale function.

When we turn off the auto.scale function, we need to determine the min and max values of the X-axis and Y-axis. 
In the previous version, the digits of the range values are 3 while the built-value digits are 1.  This cause weekness in display.

This commit is just to unify the digits